Added a Jamendo plugin for stream search
Right now it uses a client id key I created in purpose. I think this key will quickly be over used as we do a search each time we add a song to the queue and there is a 35K requests per month usage limit.
We might consider an option to use personal API keys or OAuth as it is done for lastfm scrobbling.