updated code to better match existing format
Created new actions and reducers for import related stuff. This is to better match the code with the existing format.
Moved the importing favorites logic from the hanlder in the settings to it's own files.
Added the necessary lines to the settings(container), settings(index), and App.js files to include the new action and reducer.