Fix for track-recognition error breaking library scan
Added a fix for the issue mentioned here: https://github.com/nukeop/nuclear/issues/720#issuecomment-689836722

This pull request adds a try-catch to the code processing the metadata returned by acoustic-id for a set of tracks.

Previously, if an error was encountered during this metadata processing (eg. "Empty fingerprint", as I hit), it would cause the entire local-library scanning process to fail. This was terrible for the user-experience, as it made Nuclear as a whole unusable, even if only a couple files in the library had the track-recognition issue.