Fix to handle acoustid-id result with no artist field
The acoustic-id service can apparently sometimes return a result with the track name, but no artist field. This pull-request fixes the result-handling code, to account for this possibility without erroring. (it now just stores "unknown" as the artist name, if this case is hit)

This fix should hopefully also resolve the library-scanning fail issue seen here: https://github.com/nukeop/nuclear/issues/720

(Though it's possible the issue that user hit was a different one. Either way, the issue above is fixed by this pull-request, and it's one I've hit myself.)