Remove duplicate entries for Gentoo GURU
We don't need two identical entries there for a single package.
In addition, sad news: scardracs has resigned from Gentoo in general and the package is now unmaintained. It should still keep working in the foreseeable future with minimal changes, though, as long as there is no major change in the way this program is meant to be built.