Fix Bandcamp track search caused by imprecise query
When the name of a track was very simple, the actual search match could
be on a page higher than the configured maximum. This caused the search
to fail and the track wouldn't be played. The amount of search results
can be greatly reduced by simply adding the artist name to the query.